requirements.txt
streamlit
google-cloud-firestore
seaborn